Enable adding new Senses (#14)
* Implement a function to add a new sense.

This is mostly based on wikidata-senses. Only adding senses in the same language as the lexeme is possible. Still not sure what to do about the fact that there's a delay before saving an edit and the update being fetched via the SPARQL endpoint. Perhaps getting (just) the senses directly from EntityData would be better?

Partially addresses #4. Still not possible to edit existing senses or remove them.

* Add a filter in browser.

We can now filter out lexemes that lack a sense in their own language, or that do have one.

Resolves #12.